remove flow-maven-plugin dependency from fusion-endpoint
