+++ flow-maven-plugin/src/test/java/com/vaadin/flow/plugin/maven/PrepareFrontendMojoTest.java
@@ -21,7 +21,6 @@ import org.mockito.Mockito;
 
 import com.vaadin.flow.plugin.TestUtils;
 import com.vaadin.flow.server.Constants;
-// import com.vaadin.flow.server.connect.Endpoint;
 
 import elemental.json.Json;
 import elemental.json.JsonObject;
@@ -234,13 +233,4 @@ public class PrepareFrontendMojoTest {
                 "copy-webpack-plugin", "html-webpack-plugin");
     }
 
-    // @Endpoint
-    public class MyEndpoint {
-        public void foo(String bar) {
-        }
-
-        public String bar(String baz) {
-            return baz;
-        }
-    }
 }
